+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The project uses semantic versioning (see [semver](https://semver.org)).
 
 ## [Unreleased]
 
+### Added
+
+- New `SingleFileEditor`, derived from `PatchDriver`, for editing single files.
+  - Parameter `target_file`, will be fed its text content to `process_file()`.
+  - Use via `from mass_driver.drivers.bricks import SingleFileEditor`.
+
 ## v0.17.1 - 2023-11-13
 
 ### Added